requirements.txt
django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
rest_framework_simplejwt==0.0.2
djangorestframework-simplejwt==5.2.2
